refactor: altera a estrutura do projeto para ter navegação e adiciona landing page
